Consolidate process method
